discountservice/requirements.txt
fdk>=0.1.61
mysql-connector-python







